HTTP-42-BatchRequest - add support for batch request processing in HTTP sink.
