+++ src/main/java/com/getindata/connectors/http/internal/sink/httpclient/RequestSubmitter.java
@@ -1,0 +1,13 @@
+package com.getindata.connectors.http.internal.sink.httpclient;
+
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
+
+import com.getindata.connectors.http.internal.sink.HttpSinkRequestEntry;
+
+public interface RequestSubmitter {
+
+    List<CompletableFuture<JavaNetHttpResponseWrapper>> submit(
+        String endpointUrl,
+        List<HttpSinkRequestEntry> requestToSubmit);
+}
